# api/swagger/swagger.yaml
basePath: /api/1.0/
definitions:
...
  github_com_openinfradev_tks-api_pkg_domain.FilterResponse:
    properties:
      column:
        type: string
      operator:
        type: string
      releation:
        type: string
      values:
        items:
          type: string
        type: array
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.GetSystemNotificationResponse:
    properties:
      systemNotification:
        $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SystemNotificationResponse'
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.GetSystemNotificationsResponse:
    properties:
      pagination:
        $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.PaginationResponse'
      systemNotifications:
        items:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SystemNotificationResponse'
        type: array
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.PaginationResponse:
    properties:
      filters:
        items:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.FilterResponse'
        type: array
      pageNumber:
        type: integer
      pageSize:
        type: integer
      sortColumn:
        type: string
      sortOrder:
        type: string
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.SimpleClusterResponse:
    properties:
      id:
        type: string
      name:
        type: string
      organizationId:
        type: string
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.SimpleUserResponse:
    properties:
      accountId:
        type: string
      department:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.SystemNotificationActionResponse:
    properties:
      content:
        type: string
      createdAt:
        type: string
      id:
        type: string
      status:
        type: string
      systemNotificationId:
        type: string
      taker:
        $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SimpleUserResponse'
      updatedAt:
        type: string
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.SystemNotificationResponse:
    properties:
      closedAt:
        type: string
      cluster:
        $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SimpleClusterResponse'
      createdAt:
        type: string
      firedAt:
        type: string
      grafanaUrl:
        type: string
      id:
        type: string
      lastTaker:
        $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SimpleUserResponse'
      messageActionProposal:
        type: string
      messageContent:
        type: string
      messageTitle:
        type: string
      name:
        type: string
      node:
        type: string
      notificationType:
        type: string
      organizationId:
        type: string
      policyName:
        type: string
      processingSec:
        type: integer
      rawData:
        type: string
      read:
        type: boolean
      severity:
        type: string
      status:
        type: string
      systemNotificationActions:
        items:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SystemNotificationActionResponse'
        type: array
      takedAt:
        type: string
      takedSec:
        type: integer
      updatedAt:
        type: string
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.UpdateSystemNotificationRequest:
    properties:
      description:
        type: string
    type: object
...
host: tks-api-dev.taco-cat.xyz
info:
  contact:
    email: taekyu.kang@sk.com
    name: taekyu.kang@sk.com
  description: This is backend api service for tks platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: tks-api service
  version: "1.0"
paths:
...
  /organizations/{organizationId}/system-notifications:
    get:
      consumes:
      - application/json
      description: Get SystemNotifications
      parameters:
      - description: organizationId
        in: path
        name: organizationId
        required: true
        type: string
      - description: pageSize
        in: query
        name: pageSize
        type: string
      - description: pageNumber
        in: query
        name: pageNumber
        type: string
      - description: sortColumn
        in: query
        name: soertColumn
        type: string
      - description: sortOrder
        in: query
        name: sortOrder
        type: string
      - collectionFormat: csv
        description: filters
        in: query
        items:
          type: string
        name: filters
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.GetSystemNotificationsResponse'
      security:
      - JWT: []
      summary: Get SystemNotifications
      tags:
      - SystemNotifications
  /organizations/{organizationId}/system-notifications/{systemNotificationId}:
    delete:
      consumes:
      - application/json
      description: Delete SystemNotification
      parameters:
      - description: organizationId
        in: path
        name: organizationId
        required: true
        type: string
      - description: systemNotificationId
        in: path
        name: systemNotificationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - JWT: []
      summary: Delete SystemNotification
      tags:
      - SystemNotifications
    get:
      consumes:
      - application/json
      description: Get SystemNotification
      parameters:
      - description: organizationId
        in: path
        name: organizationId
        required: true
        type: string
      - description: systemNotificationId
        in: path
        name: systemNotificationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.GetSystemNotificationResponse'
      security:
      - JWT: []
      summary: Get SystemNotification
      tags:
      - SystemNotifications
    put:
      consumes:
      - application/json
      description: Update SystemNotification
      parameters:
      - description: organizationId
        in: path
        name: organizationId
        required: true
        type: string
      - description: systemNotificationId
        in: path
        name: systemNotificationId
        required: true
        type: string
      - description: Update cloud setting request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.UpdateSystemNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - JWT: []
      summary: Update SystemNotification
      tags:
      - SystemNotifications
  /organizations/{organizationId}/system-notifications/{systemNotificationId}/actions:
    post:
      consumes:
      - application/json
      description: Create systemNotification action
      parameters:
      - description: organizationId
        in: path
        name: organizationId
        required: true
        type: string
      - description: systemNotificationId
        in: path
        name: systemNotificationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - JWT: []
      summary: Create systemNotification action
      tags:
      - SystemNotifications
...
  /system-api/organizations/{organizationId}/system-notifications:
    post:
      consumes:
      - application/json
      description: Create systemNotification. ADMIN ONLY
      parameters:
      - description: organizationId
        in: path
        name: organizationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - JWT: []
      summary: Create systemNotification. ADMIN ONLY
      tags:
      - SystemNotifications
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"